refactor: extract resumes-diagnostics.ts from resumes.ts
Move diagnostics types (IngestDiagnosticsRow, DiagnosticsSourceFacetRow,
IngestDiagnosticsBrandHit, IngestDiagnosticsTaggingEntry) and helper
functions (resolveDiagnosticsSourceKeyForResume, matchesDiagnosticsSourceKeys,
buildDiagnosticsSourceFacetRows, projectIngestDiagnosticsRow,
normalizeDiagnosticsSourceFilterValues) to lib/resumes_diagnostics.ts.

resumes.ts re-exports all moved symbols for backward compatibility.
resumes.ts reduced from 3365 to 3195 lines (5%).
